move error to resource
